Refacoring codee as PR comments
- Moved resumable signed_url from Resource -> File model.
- Add utility function to convert string to boolean.
- Added and refactored unit tests.
